phpredis: move each version tested to its own folder
